Updates - Draggable mid, longer body & collapse about
Implemeted the core structure for creating custom scrolls.
Still need to get it working.

Made the body 2400px ~ longer. Will need to find a better way
of doing that. The middle orange div is now actually draggable.
There is a collapsable panel in the top left.
Might not be the best option but it works. Will fix and spruce
that up later too.